beam_pipelines/requirements.txt
apache-beam[gcp,dataframe]~=2.51.0
Pillow~=10.0.1
tensorflow~=2.15.0
